requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
PyYAML>=5.4.0
tqdm>=4.62.0
Pillow>=8.3.0
albumentations>=1.3.0

hydra-core>=1.3.2
